Split tasks graph
