+++ ydb/core/kqp/executer_actor/kqp_table_resolver.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "kqp_tasks_graph.h"
+#include "tasks_graph/kqp_tasks_graph.h"
 
 namespace NKikimr::NKqp {
