--- ydb/core/kqp/executer_actor/kqp_tasks_validate.h
@@ -1,12 +1,0 @@
-#pragma once
-
-#include "kqp_executer_impl.h"
-#include "kqp_tasks_graph.h"
-
-namespace NKikimr {
-namespace NKqp {
-
-bool ValidateTasks(const TKqpTasksGraph& tasksGraph, const EExecType& execType, bool enableSpilling, NYql::TIssue& issue);
-
-} // namespace NKqp
-} // namespace NKikimr
