+++ ydb/core/kqp/executer_actor/tasks_graph/kqp_tasks_graph.cpp
@@ -38,12 +38,7 @@ using namespace NYql::NNodes;
 
 namespace {
 
-struct TStageScheduleInfo {
-    double StageCost = 0.0;
-    ui32 TaskCount = 0;
-};
-
-void ParseColumnToProto(TString columnName,
+void ParseColumnToProto(const TString& columnName,
     TMap<TString, NSharding::IShardingBase::TColumn>::const_iterator columnIt,
     ::NKikimrKqp::TKqpColumnMetadataProto* columnProto)
 {
@@ -56,229 +51,14 @@ void ParseColumnToProto(TString columnName,
     }
 };
 
-std::map<ui32, TStageScheduleInfo> ScheduleByCost(const IKqpGateway::TPhysicalTxData& tx, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot) {
-    std::map<ui32, TStageScheduleInfo> result;
-    if (!resourceSnapshot.empty()) { // can't schedule w/o node count
-        // collect costs and schedule stages with external sources only
-        double totalCost = 0.0;
-        for (ui32 stageIdx = 0; stageIdx < tx.Body->StagesSize(); ++stageIdx) {
-            auto& stage = tx.Body->GetStages(stageIdx);
-            if (stage.SourcesSize() > 0 && stage.GetSources(0).GetTypeCase() == NKqpProto::TKqpSource::kExternalSource) {
-                if (stage.GetStageCost() > 0.0 && stage.GetTaskCount() == 0) {
-                    totalCost += stage.GetStageCost();
-                    result.emplace(stageIdx, TStageScheduleInfo{.StageCost = stage.GetStageCost()});
-                }
-            }
-        }
-        // assign task counts
-        if (!result.empty()) {
-            // allow use 2/3 of threads in single stage
-            ui32 maxStageTaskCount = (TStagePredictor::GetUsableThreads() * 2 + 2) / 3;
-            // total limit per mode is x2
-            ui32 maxTotalTaskCount = maxStageTaskCount * 2;
-            for (auto& [_, stageInfo] : result) {
-                // schedule tasks evenly between nodes
-                stageInfo.TaskCount =
-                    std::max<ui32>(
-                        std::min(static_cast<ui32>(maxTotalTaskCount * stageInfo.StageCost / totalCost), maxStageTaskCount)
-                        , 1
-                    ) * resourceSnapshot.size();
-            }
-        }
-    }
-    return result;
-}
-
-void FillReadInfo(TTaskMeta& taskMeta, ui64 itemsLimit, const NYql::ERequestSorting sorting) {
-    if (taskMeta.Reads && !taskMeta.Reads.GetRef().empty()) {
-        // Validate parameters
-        YQL_ENSURE(taskMeta.ReadInfo.ItemsLimit == itemsLimit);
-        YQL_ENSURE(taskMeta.ReadInfo.GetSorting() == sorting);
-        // TODO: why no check for ReadType?
-        return;
-    }
-
-    taskMeta.ReadInfo.ItemsLimit = itemsLimit;
-    taskMeta.ReadInfo.SetSorting(sorting);
-    taskMeta.ReadInfo.ReadType = TTaskMeta::TReadInfo::EReadType::Rows;
-}
-
-TTaskMeta::TReadInfo::EReadType OlapReadTypeFromProto(const NKqpProto::TKqpPhyOpReadOlapRanges::EReadType& type) {
-    switch (type) {
-        case NKqpProto::TKqpPhyOpReadOlapRanges::ROWS:
-            return TTaskMeta::TReadInfo::EReadType::Rows;
-        case NKqpProto::TKqpPhyOpReadOlapRanges::BLOCKS:
-            return TTaskMeta::TReadInfo::EReadType::Blocks;
-        default:
-            YQL_ENSURE(false, "Invalid read type from TKqpPhyOpReadOlapRanges protobuf.");
-    }
-}
-
-void FillOlapReadInfo(TTaskMeta& taskMeta, NKikimr::NMiniKQL::TType* resultType, const TMaybe<::NKqpProto::TKqpPhyOpReadOlapRanges>& readOlapRange) {
-    if (taskMeta.Reads && !taskMeta.Reads.GetRef().empty()) {
-        // Validate parameters
-        if (!readOlapRange || readOlapRange->GetOlapProgram().empty()) {
-            YQL_ENSURE(taskMeta.ReadInfo.OlapProgram.Program.empty());
-            return;
-        }
-
-        YQL_ENSURE(taskMeta.ReadInfo.OlapProgram.Program == readOlapRange->GetOlapProgram());
-        return;
-    }
-
-    if (resultType) {
-        YQL_ENSURE(resultType->GetKind() == NKikimr::NMiniKQL::TType::EKind::Struct
-            || resultType->GetKind() == NKikimr::NMiniKQL::TType::EKind::Tuple);
-
-        auto* resultStructType = static_cast<NKikimr::NMiniKQL::TStructType*>(resultType);
-        ui32 resultColsCount = resultStructType->GetMembersCount();
-
-        taskMeta.ReadInfo.ResultColumnsTypes.reserve(resultColsCount);
-        for (ui32 i = 0; i < resultColsCount; ++i) {
-            taskMeta.ReadInfo.ResultColumnsTypes.emplace_back();
-            auto memberType = resultStructType->GetMemberType(i);
-            NScheme::TTypeInfo typeInfo = NScheme::TypeInfoFromMiniKQLType(memberType);
-            taskMeta.ReadInfo.ResultColumnsTypes.back() = typeInfo;
-        }
-    }
-    if (!readOlapRange || readOlapRange->GetOlapProgram().empty()) {
-        return;
-    }
-    {
-        Y_ABORT_UNLESS(taskMeta.ReadInfo.GroupByColumnNames.empty());
-        std::vector<std::string> groupByColumns;
-        for (auto&& i : readOlapRange->GetGroupByColumnNames()) {
-            groupByColumns.emplace_back(i);
-        }
-        std::swap(taskMeta.ReadInfo.GroupByColumnNames, groupByColumns);
-    }
-    taskMeta.ReadInfo.ReadType = OlapReadTypeFromProto(readOlapRange->GetReadType());
-    taskMeta.ReadInfo.OlapProgram.Program = readOlapRange->GetOlapProgram();
-    for (auto& name: readOlapRange->GetOlapProgramParameterNames()) {
-        taskMeta.ReadInfo.OlapProgram.ParameterNames.insert(name);
-    }
-}
-
-void MergeReadInfoToTaskMeta(TTaskMeta& meta, ui64 shardId, TMaybe<TShardKeyRanges>& keyReadRanges,
-    const TPhysicalShardReadSettings& readSettings, const TVector<TTaskMeta::TColumn>& columns,
-    const NKqpProto::TKqpPhyTableOperation& op, bool isPersistentScan)
-{
-    TTaskMeta::TShardReadInfo readInfo = {
-        .Ranges = {},
-        .Columns = columns,
-    };
-    if (keyReadRanges) {
-        readInfo.Ranges = std::move(*keyReadRanges); // sorted & non-intersecting
-    }
-
-    if (isPersistentScan) {
-        readInfo.ShardId = shardId;
-    }
-
-    FillReadInfo(meta, readSettings.ItemsLimit, readSettings.GetSorting());
-    if (op.GetTypeCase() == NKqpProto::TKqpPhyTableOperation::kReadOlapRange) {
-        FillOlapReadInfo(meta, readSettings.ResultType, op.GetReadOlapRange());
-    }
-
-    if (!meta.Reads) {
-        meta.Reads.ConstructInPlace();
-    }
-
-    meta.Reads->emplace_back(std::move(readInfo));
-}
-
-void PrepareScanMetaForUsage(TTaskMeta& meta, const TVector<NScheme::TTypeInfo>& keyTypes) {
-    YQL_ENSURE(meta.Reads.Defined());
-    auto& taskReads = meta.Reads.GetRef();
-
-    /*
-     * Sort read ranges so that sequential scan of that ranges produce sorted result.
-     *
-     * Partition pruner feed us with set of non-intersecting ranges with filled right boundary.
-     * So we may sort ranges based solely on the their rightmost point.
-     */
-    std::sort(taskReads.begin(), taskReads.end(), [&](const auto& lhs, const auto& rhs) {
-        if (lhs.ShardId == rhs.ShardId) {
-            return false;
-        }
-
-        const std::pair<const TSerializedCellVec*, bool> k1 = lhs.Ranges.GetRightBorder();
-        const std::pair<const TSerializedCellVec*, bool> k2 = rhs.Ranges.GetRightBorder();
-
-        const int cmp = CompareBorders<false, false>(
-            k1.first->GetCells(),
-            k2.first->GetCells(),
-            k1.second,
-            k2.second,
-            keyTypes);
-
-        return (cmp < 0);
-        });
-}
-
-void FillReadTaskFromSource(TTask& task, const TString& sourceName, const TString& structuredToken, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot, ui64 nodeOffset) {
-    if (structuredToken) {
-        task.Meta.SecureParams.emplace(sourceName, structuredToken);
-    }
-
-    if (resourceSnapshot.empty()) {
-        task.Meta.Type = TTaskMeta::TTaskType::Compute;
-    } else {
-        task.Meta.NodeId = resourceSnapshot[nodeOffset % resourceSnapshot.size()].GetNodeId();
-        task.Meta.Type = TTaskMeta::TTaskType::Scan;
-    }
-}
-
-struct TShardRangesWithShardId {
-    TMaybe<ui64> ShardId;
-    const TShardKeyRanges* Ranges;
-};
-
-TVector<TVector<TShardRangesWithShardId>> DistributeShardsToTasks(TVector<TShardRangesWithShardId> shardsRanges, const size_t tasksCount, const TVector<NScheme::TTypeInfo>& keyTypes) {
-    // TODO:
-    // if (IsDebugLogEnabled()) {
-    //     TStringBuilder sb;
-    //     sb << "Distributing shards to tasks: [";
-    //     for(size_t i = 0; i < shardsRanges.size(); i++) {
-    //         sb << "# " << i << ": " << shardsRanges[i].Ranges->ToString(keyTypes, *AppData()->TypeRegistry);
-    //     }
-
-    //     sb << " ].";
-    //     LOG_D(sb);
-    // }
-
-    std::sort(std::begin(shardsRanges), std::end(shardsRanges), [&](const TShardRangesWithShardId& lhs, const TShardRangesWithShardId& rhs) {
-            return CompareBorders<true, true>(
-                lhs.Ranges->GetRightBorder().first->GetCells(),
-                rhs.Ranges->GetRightBorder().first->GetCells(),
-                lhs.Ranges->GetRightBorder().second,
-                rhs.Ranges->GetRightBorder().second,
-                keyTypes) < 0;
-        });
-
-    // One shard (ranges set) can be assigned only to one task. Otherwise, we can break some optimizations like removing unnecessary shuffle.
-    TVector<TVector<TShardRangesWithShardId>> result(tasksCount);
-    size_t shardIndex = 0;
-    for (size_t taskIndex = 0; taskIndex < tasksCount; ++taskIndex) {
-        const size_t tasksLeft = tasksCount - taskIndex;
-        const size_t shardsLeft = shardsRanges.size() - shardIndex;
-        const size_t shardsPerCurrentTask = (shardsLeft + tasksLeft - 1) / tasksLeft;
-
-        for (size_t currentShardIndex = 0; currentShardIndex < shardsPerCurrentTask; ++currentShardIndex, ++shardIndex) {
-            result[taskIndex].push_back(shardsRanges[shardIndex]);
-        }
-    }
-    return result;
-}
-
 template<typename Proto>
-TVector<TTaskMeta::TColumn> BuildKqpColumns(const Proto& op, TIntrusiveConstPtr<TTableConstInfo> tableInfo) {
+TVector<TTaskMeta::TColumn> BuildKqpColumnsImpl(const Proto& op, const TIntrusiveConstPtr<TTableConstInfo>& tableInfo) {
     TVector<TTaskMeta::TColumn> columns;
     columns.reserve(op.GetColumns().size());
 
     THashSet<TString> keyColumns;
     for (auto column : tableInfo->KeyColumns) {
-        keyColumns.insert(std::move(column));
+        keyColumns.emplace(std::move(column));
     }
 
     for (const auto& column : op.GetColumns()) {
@@ -299,7 +79,7 @@ TVector<TTaskMeta::TColumn> BuildKqpColumns(const Proto& op, TIntrusiveConstPtr<
 }
 
 struct TKqpTaskOutputType {
-    enum : ui32 {
+    enum : ui8 {
         ShardRangePartition = TTaskOutputType::COMMON_TASK_OUTPUT_TYPE_END
     };
 };
@@ -429,7 +209,7 @@ void FillTaskMeta(const TStageInfo& stageInfo, const TTask& task, NYql::NDqProto
                 protoTaskMeta.AddGroupByColumnNames(i.data(), i.size());
             }
 
-            for (auto columnType : task.Meta.ReadInfo.ResultColumnsTypes) {
+            for (const auto& columnType : task.Meta.ReadInfo.ResultColumnsTypes) {
                 auto* protoResultColumn = protoTaskMeta.AddResultColumns();
                 protoResultColumn->SetId(0);
                 auto protoColumnType = NScheme::ProtoColumnTypeFromTypeInfoMod(columnType, "");
@@ -451,7 +231,7 @@ void FillTaskMeta(const TStageInfo& stageInfo, const TTask& task, NYql::NDqProto
                 YQL_ENSURE(task.Meta.ReadInfo.OlapProgram.Program.empty());
             }
 
-            for (auto& column : task.Meta.Reads->front().Columns) {
+            for (const auto& column : task.Meta.Reads->front().Columns) {
                 auto* protoColumn = protoTaskMeta.AddColumns();
                 protoColumn->SetId(column.Id);
                 auto columnType = NScheme::ProtoColumnTypeFromTypeInfoMod(column.Type, "");
@@ -463,7 +243,7 @@ void FillTaskMeta(const TStageInfo& stageInfo, const TTask& task, NYql::NDqProto
             }
         }
 
-        for (auto& read : *task.Meta.Reads) {
+        for (const auto& read : *task.Meta.Reads) {
             auto* protoReadMeta = protoTaskMeta.AddReads();
             protoReadMeta->SetShardId(read.ShardId);
             read.Ranges.SerializeTo(protoReadMeta);
@@ -519,7 +299,7 @@ void TKqpTasksGraph::FillKqpTasksGraphStages() {
                         meta.TableId = MakeTableId(source.GetFullTextSource().GetTable());
                         meta.TablePath = source.GetFullTextSource().GetTable().GetPath();
                         meta.ShardOperations.insert(TKeyDesc::ERowOperation::Read);
-                        YQL_ENSURE(tx.Body->GetTableConstInfoById()->Map.find(meta.TableId) != tx.Body->GetTableConstInfoById()->Map.end(), "Cannot find table const info for table: " << meta.TableId);
+                        YQL_ENSURE(tx.Body->GetTableConstInfoById()->Map.contains(meta.TableId), "Cannot find table const info for table: " << meta.TableId);
                         meta.TableConstInfo = tx.Body->GetTableConstInfoById()->Map.at(meta.TableId);
                         stageSourcesCount++;
                         break;
@@ -595,7 +375,7 @@ void TKqpTasksGraph::FillKqpTasksGraphStages() {
                 }
             };
 
-            for (auto& sink : stage.GetSinks()) {
+            for (const auto& sink : stage.GetSinks()) {
                 if (sink.GetTypeCase() == NKqpProto::TKqpSink::kInternalSink && sink.GetInternalSink().GetSettings().Is<NKikimrKqp::TKqpTableSinkSettings>()) {
                     NKikimrKqp::TKqpTableSinkSettings settings;
                     YQL_ENSURE(sink.GetInternalSink().GetSettings().UnpackTo(&settings), "Failed to unpack settings");
@@ -605,7 +385,7 @@ void TKqpTasksGraph::FillKqpTasksGraphStages() {
                 }
             }
 
-            for (auto& transform : stage.GetOutputTransforms()) {
+            for (const auto& transform : stage.GetOutputTransforms()) {
                 if (transform.GetTypeCase() == NKqpProto::TKqpOutputTransform::kInternalSink && transform.GetInternalSink().GetSettings().Is<NKikimrKqp::TKqpTableSinkSettings>()) {
                     NKikimrKqp::TKqpTableSinkSettings settings;
                     YQL_ENSURE(transform.GetInternalSink().GetSettings().UnpackTo(&settings), "Failed to unpack settings");
@@ -646,7 +426,7 @@ void TKqpTasksGraph::FillKqpTasksGraphStages() {
                 }
             }
 
-            YQL_ENSURE(tables.empty() || tables.size() == 1);
+            YQL_ENSURE(tables.size() <= 1);
         }
     }
 }
@@ -719,7 +499,7 @@ void TKqpTasksGraph::BuildSequencerChannels(const TStageInfo& stageInfo, ui32 in
 {
     YQL_ENSURE(stageInfo.Tasks.size() == inputStageInfo.Tasks.size());
 
-    NKikimrKqp::TKqpSequencerSettings* settings = GetMeta().Allocate<NKikimrKqp::TKqpSequencerSettings>();
+    auto* settings = GetMeta().Allocate<NKikimrKqp::TKqpSequencerSettings>();
     settings->MutableTable()->CopyFrom(sequencer.GetTable());
     settings->SetDatabase(GetMeta().Database);
     settings->MutableColumns()->CopyFrom(sequencer.GetColumns());
@@ -779,7 +559,7 @@ void TKqpTasksGraph::BuildStreamLookupChannels(const TStageInfo& stageInfo, ui32
 {
     YQL_ENSURE(stageInfo.Tasks.size() == inputStageInfo.Tasks.size());
 
-    NKikimrKqp::TKqpStreamLookupSettings* settings = GetMeta().Allocate<NKikimrKqp::TKqpStreamLookupSettings>();
+    auto* settings = GetMeta().Allocate<NKikimrKqp::TKqpStreamLookupSettings>();
 
     settings->SetDatabase(GetMeta().Database);
     settings->MutableTable()->CopyFrom(streamLookup.GetTable());
@@ -827,7 +607,7 @@ void TKqpTasksGraph::BuildStreamLookupChannels(const TStageInfo& stageInfo, ui32
         auto target = ExtractPhyValue(stageInfo, in.GetTargetVector(), TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod());
         out.SetTargetVector(TString(target.AsStringRef()));
         out.SetLimit((ui32)ExtractPhyValue(stageInfo, in.GetLimit(), TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod()).Get<ui64>());
-        for (auto& colIdx: in.GetDistinctColumns()) {
+        for (const auto& colIdx: in.GetDistinctColumns()) {
             out.AddDistinctColumns(colIdx);
         }
     }
@@ -858,7 +638,7 @@ void TKqpTasksGraph::BuildVectorResolveChannels(const TStageInfo& stageInfo, ui3
 
     settings->SetDatabase(GetMeta().Database);
     auto* levelMeta = settings->MutableLevelTable();
-    auto& kqpMeta = vectorResolve.GetLevelTable();
+    const auto& kqpMeta = vectorResolve.GetLevelTable();
     levelMeta->SetTablePath(kqpMeta.GetPath());
     levelMeta->MutableTableId()->SetTableId(kqpMeta.GetTableId());
     levelMeta->MutableTableId()->SetOwnerId(kqpMeta.GetOwnerId());
@@ -961,7 +741,7 @@ void TKqpTasksGraph::BuildDqSourceStreamLookupChannels(const TStageInfo& stageIn
 }
 
 void TKqpTasksGraph::BuildKqpStageChannels(TStageInfo& stageInfo, ui64 txId, bool enableSpilling, bool enableShuffleElimination) {
-    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
+    const auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
 
     if (stage.GetIsEffectsStage() && stage.GetSinks().empty()) {
         YQL_ENSURE(stageInfo.OutputsCount == 1);
@@ -1014,7 +794,7 @@ void TKqpTasksGraph::BuildKqpStageChannels(TStageInfo& stageInfo, ui64 txId, boo
             (*columnShardHashV1Params.TaskIndexByHash)[i] = i;
         }
 
-        for (auto& input : stage.GetInputs()) {
+        for (const auto& input : stage.GetInputs()) {
             if (input.GetTypeCase() != NKqpProto::TKqpPhyConnection::kHashShuffle) {
                 continue;
             }
@@ -1028,7 +808,7 @@ void TKqpTasksGraph::BuildKqpStageChannels(TStageInfo& stageInfo, ui64 txId, boo
             // ^ if the flag if false, and kColumnShardHashV1 detected - then the data which would be returned - would be incorrect,
             // because we didn't save partitioning in the BuildScanTasksFromShards.
 
-            auto columnShardHashV1 = hashShuffle.GetColumnShardHashV1();
+            const auto& columnShardHashV1 = hashShuffle.GetColumnShardHashV1();
             columnShardHashV1Params.SourceTableKeyColumnTypes = std::make_shared<TVector<NScheme::TTypeInfo>>();
             columnShardHashV1Params.SourceTableKeyColumnTypes->reserve(columnShardHashV1.KeyColumnTypesSize());
             for (const auto& keyColumnType: columnShardHashV1.GetKeyColumnTypes()) {
@@ -1042,7 +822,7 @@ void TKqpTasksGraph::BuildKqpStageChannels(TStageInfo& stageInfo, ui64 txId, boo
     }
 
     ui64 nextOriginTaskId = 0;
-    for (auto& input : stage.GetInputs()) {
+    for (const auto& input : stage.GetInputs()) {
         ui32 inputIdx = input.GetInputIndex();
         auto& inputStageInfo = GetStageInfo(TStageId(stageInfo.Id.TxId, input.GetStageIndex()));
         const auto& outputIdx = input.GetOutputIndex();
@@ -1159,7 +939,7 @@ void TKqpTasksGraph::BuildKqpStageChannels(TStageInfo& stageInfo, ui64 txId, boo
 }
 
 void TKqpTasksGraph::FillChannelDesc(NDqProto::TChannel& channelDesc, const TChannel& channel,
-    const NKikimrConfig::TTableServiceConfig::EChannelTransportVersion chanTransportVersion, bool enableSpilling) const
+    const NKikimrConfig::TTableServiceConfig::EChannelTransportVersion channelTransportVersion, bool enableSpilling) const
 {
     channelDesc.SetId(channel.Id);
     channelDesc.SetSrcStageId(channel.SrcStageId.StageId);
@@ -1193,7 +973,7 @@ void TKqpTasksGraph::FillChannelDesc(NDqProto::TChannel& channelDesc, const TCha
 
     channelDesc.SetIsPersistent(false);
     channelDesc.SetInMemory(channel.InMemory);
-    if (chanTransportVersion == NKikimrConfig::TTableServiceConfig::CTV_OOB_PICKLE_1_0) {
+    if (channelTransportVersion == NKikimrConfig::TTableServiceConfig::CTV_OOB_PICKLE_1_0) {
         channelDesc.SetTransportVersion(NDqProto::EDataTransportVersion::DATA_TRANSPORT_OOB_PICKLE_1_0);
     } else {
         channelDesc.SetTransportVersion(NDqProto::EDataTransportVersion::DATA_TRANSPORT_UV_PICKLE_1_0);
@@ -1211,7 +991,7 @@ void TKqpTasksGraph::FillOutputDesc(NYql::NDqProto::TTaskOutput& outputDesc, con
 
         case TTaskOutputType::HashPartition: {
             auto& hashPartitionDesc = *outputDesc.MutableHashPartition();
-            for (auto& column : output.KeyColumns) {
+            for (const auto& column : output.KeyColumns) {
                 hashPartitionDesc.AddKeyColumns(column);
             }
             hashPartitionDesc.SetPartitionsCount(output.PartitionsCount);
@@ -1229,7 +1009,7 @@ void TKqpTasksGraph::FillOutputDesc(NYql::NDqProto::TTaskOutput& outputDesc, con
                     break;
                 }
                 case ColumnShardHashV1: {
-                    auto& columnShardHashV1Params = stageInfo.Meta.GetColumnShardHashV1Params(outputIdx);
+                    const auto& columnShardHashV1Params = stageInfo.Meta.GetColumnShardHashV1Params(outputIdx);
                     LOG_D( "Filling columnshardhashv1 params for sending it to runtime "
                         << "[" << stageInfo.Id.TxId << ":" << stageInfo.Id.StageId << "]"
                         << ": " << columnShardHashV1Params.KeyTypesToString()
@@ -1268,13 +1048,13 @@ void TKqpTasksGraph::FillOutputDesc(NYql::NDqProto::TTaskOutput& outputDesc, con
         case TKqpTaskOutputType::ShardRangePartition: {
             auto& rangePartitionDesc = *outputDesc.MutableRangePartition();
             auto& columns = *rangePartitionDesc.MutableKeyColumns();
-            for (auto& column : output.KeyColumns) {
+            for (const auto& column : output.KeyColumns) {
                 *columns.Add() = column;
             }
 
             auto& partitionsDesc = *rangePartitionDesc.MutablePartitions();
-            for (auto& pair : output.Meta.ShardPartitions) {
-                auto& range = *pair.second->Range;
+            for (const auto& pair : output.Meta.ShardPartitions) {
+                const auto& range = *pair.second->Range;
                 auto& partitionDesc = *partitionsDesc.Add();
                 partitionDesc.SetEndKeyPrefix(range.EndKeyPrefix.GetBuffer());
                 partitionDesc.SetIsInclusive(range.IsInclusive);
@@ -1307,7 +1087,7 @@ void TKqpTasksGraph::FillOutputDesc(NYql::NDqProto::TTaskOutput& outputDesc, con
         }
     }
 
-    for (auto& channel : output.Channels) {
+    for (const auto& channel : output.Channels) {
         auto& channelDesc = *outputDesc.AddChannels();
         FillChannelDesc(channelDesc, GetChannel(channel), GetMeta().ChannelTransportVersion, enableSpilling);
     }
@@ -1379,9 +1159,9 @@ void TKqpTasksGraph::FillInputDesc(NYql::NDqProto::TTaskInput& inputDesc, const 
         case NYql::NDq::TTaskInputType::Merge: {
             auto& mergeProto = *inputDesc.MutableMerge();
             YQL_ENSURE(std::holds_alternative<NYql::NDq::TMergeTaskInput>(input.ConnectionInfo));
-            auto& sortColumns = std::get<NYql::NDq::TMergeTaskInput>(input.ConnectionInfo).SortColumns;
+            const auto& sortColumns = std::get<NYql::NDq::TMergeTaskInput>(input.ConnectionInfo).SortColumns;
             for (const auto& sortColumn : sortColumns) {
-                auto newSortCol = mergeProto.AddSortColumns();
+                auto* newSortCol = mergeProto.AddSortColumns();
                 newSortCol->SetColumn(sortColumn.Column.c_str());
                 newSortCol->SetAscending(sortColumn.Ascending);
             }
@@ -1472,7 +1252,7 @@ void TKqpTasksGraph::FillInputDesc(NYql::NDqProto::TTaskInput& inputDesc, const 
 }
 
 void TKqpTasksGraph::SerializeTaskToProto(const TTask& task, NYql::NDqProto::TDqTask* result, bool serializeAsyncIoSettings) const {
-    auto& stageInfo = GetStageInfo(task.StageId);
+    const auto& stageInfo = GetStageInfo(task.StageId);
     ActorIdToProto(task.Meta.ExecuterId, result->MutableExecuter()->MutableActorId());
     result->SetId(task.Id);
     result->SetStageId(stageInfo.Id.StageId);
@@ -1511,7 +1291,7 @@ void TKqpTasksGraph::SerializeTaskToProto(const TTask& task, NYql::NDqProto::TDq
     const NKqpProto::TKqpPhyStage& stage = stageInfo.Meta.GetStage(stageInfo.Id);
     result->MutableProgram()->CopyFrom(stage.GetProgram());
 
-    for (auto& paramName : stage.GetProgramParameters()) {
+    for (const auto& paramName : stage.GetProgramParameters()) {
         auto& dqParams = *result->MutableParameters();
         if (task.Meta.ShardId) {
             dqParams[paramName] = stageInfo.Meta.Tx.Params->GetShardParam(task.Meta.ShardId, paramName);
@@ -1534,7 +1314,7 @@ void TKqpTasksGraph::SerializeTaskToProto(const TTask& task, NYql::NDqProto::TDq
 }
 
 NYql::NDqProto::TDqTask* TKqpTasksGraph::ArenaSerializeTaskToProto(const TTask& task, bool serializeAsyncIoSettings) {
-    NYql::NDqProto::TDqTask* result = GetMeta().Allocate<NYql::NDqProto::TDqTask>();
+    auto* result = GetMeta().Allocate<NYql::NDqProto::TDqTask>();
     SerializeTaskToProto(task, result, serializeAsyncIoSettings);
     return result;
 }
@@ -1557,8 +1337,6 @@ void TKqpTasksGraph::PersistTasksGraphInfo(NKikimrKqp::TQueryPhysicalGraph& resu
 }
 
 void TKqpTasksGraph::RestoreTasksGraphInfo(const TVector<NKikimrKqp::TKqpNodeResources>& resourcesSnapshot, const NKikimrKqp::TQueryPhysicalGraph& graphInfo) {
-    GetMeta().IsRestored = true;
-
     const auto restoreDqTransform = [](const auto& protoInfo) -> TMaybe<TTransform> {
         if (!protoInfo.HasTransform()) {
             return Nothing();
@@ -1581,16 +1359,16 @@ void TKqpTasksGraph::RestoreTasksGraphInfo(const TVector<NKikimrKqp::TKqpNodeRes
 
         const auto& settings = protoInfo.GetTransform().GetSettings();
         if (settings.Is<NKikimrKqp::TKqpStreamLookupSettings>()) {
-            auto transformSettings = meta.StreamLookupSettings = GetMeta().Allocate<NKikimrKqp::TKqpStreamLookupSettings>();
+            auto* transformSettings = meta.StreamLookupSettings = GetMeta().Allocate<NKikimrKqp::TKqpStreamLookupSettings>();
             YQL_ENSURE(settings.UnpackTo(transformSettings), "Failed to parse stream lookup settings");
             transformSettings->ClearSnapshot();
             transformSettings->ClearLockTxId();
             transformSettings->ClearLockNodeId();
         } else if (settings.Is<NKikimrKqp::TKqpSequencerSettings>()) {
-            auto transformSettings = meta.SequencerSettings = GetMeta().Allocate<NKikimrKqp::TKqpSequencerSettings>();
+            auto* transformSettings = meta.SequencerSettings = GetMeta().Allocate<NKikimrKqp::TKqpSequencerSettings>();
             YQL_ENSURE(settings.UnpackTo(transformSettings), "Failed to parse sequencer settings");
         } else if (settings.Is<NKikimrTxDataShard::TKqpVectorResolveSettings>()) {
-            auto transformSettings = meta.VectorResolveSettings = GetMeta().Allocate<NKikimrTxDataShard::TKqpVectorResolveSettings>();
+            auto* transformSettings = meta.VectorResolveSettings = GetMeta().Allocate<NKikimrTxDataShard::TKqpVectorResolveSettings>();
             YQL_ENSURE(settings.UnpackTo(transformSettings), "Failed to parse vector resolve settings");
             transformSettings->ClearSnapshot();
             transformSettings->ClearLockTxId();
@@ -1634,7 +1412,7 @@ void TKqpTasksGraph::RestoreTasksGraphInfo(const TVector<NKikimrKqp::TKqpNodeRes
         newTask.SetUseLlvm(taskInfo.GetUseLlvm());
         newTask.Meta.TaskParams.insert(taskInfo.GetTaskParams().begin(), taskInfo.GetTaskParams().end());
         newTask.Meta.ReadRanges.assign(taskInfo.GetReadRanges().begin(), taskInfo.GetReadRanges().end());
-        newTask.Meta.Type = TTaskMeta::TTaskType::Compute;
+        newTask.Meta.Type = TTaskMeta::ETaskType::Compute;
         newTask.Meta.ExecuterId = GetMeta().ExecuterId;
 
         if (taskInfo.HasMetaId()) {
@@ -1725,16 +1503,16 @@ void TKqpTasksGraph::RestoreTasksGraphInfo(const TVector<NKikimrKqp::TKqpNodeRes
                     if (sourceInfo.HasSettings()) {
                         const auto& settings = sourceInfo.GetSettings();
                         if (settings.Is<NKikimrTxDataShard::TKqpReadRangesSourceSettings>()) {
-                            auto sourceSettings = newInput.Meta.SourceSettings = GetMeta().Allocate<NKikimrTxDataShard::TKqpReadRangesSourceSettings>();
+                            auto* sourceSettings = newInput.Meta.SourceSettings = GetMeta().Allocate<NKikimrTxDataShard::TKqpReadRangesSourceSettings>();
                             YQL_ENSURE(settings.UnpackTo(sourceSettings), "Failed to parse source settings");
                             FillScanTaskLockTxId(*sourceSettings);
                             sourceSettings->ClearSnapshot();
                         } else if (settings.Is<NKikimrKqp::TKqpFullTextSourceSettings>()) {
-                            auto sourceSettings = newInput.Meta.FullTextSourceSettings = GetMeta().Allocate<NKikimrKqp::TKqpFullTextSourceSettings>();
+                            auto* sourceSettings = newInput.Meta.FullTextSourceSettings = GetMeta().Allocate<NKikimrKqp::TKqpFullTextSourceSettings>();
                             YQL_ENSURE(settings.UnpackTo(sourceSettings), "Failed to parse full text source settings");
                             sourceSettings->ClearSnapshot();
                         } else if (settings.Is<NKikimrKqp::TKqpSysViewSourceSettings>()) {
-                            auto sourceSettings = newInput.Meta.SysViewSourceSettings = GetMeta().Allocate<NKikimrKqp::TKqpSysViewSourceSettings>();
+                            auto* sourceSettings = newInput.Meta.SysViewSourceSettings = GetMeta().Allocate<NKikimrKqp::TKqpSysViewSourceSettings>();
                             YQL_ENSURE(settings.UnpackTo(sourceSettings), "Failed to parse sys view source settings");
                         } else {
                             newInput.SourceSettings = sourceInfo.GetSettings();
@@ -1865,60 +1643,11 @@ void TKqpTasksGraph::RestoreTasksGraphInfo(const TVector<NKikimrKqp::TKqpNodeRes
 
             if (const auto& sources = stage.GetSources(); !sources.empty() && sources[0].GetTypeCase() == NKqpProto::TKqpSource::kExternalSource) {
                 const auto it = scheduledTaskCount.find(stageIdx);
-                BuildReadTasksFromSource(stageInfo, resourcesSnapshot, it != scheduledTaskCount.end() ? it->second.TaskCount : 0);
+                RestoreReadTasksFromSource(stageInfo, resourcesSnapshot, it != scheduledTaskCount.end() ? it->second.TaskCount : 0);
             }
 
             GetMeta().AllowWithSpilling |= stage.GetAllowWithSpilling();
         }
-    }
-}
-
-void TKqpTasksGraph::BuildSysViewScanTasks(TStageInfo& stageInfo) {
-    Y_DEBUG_ABORT_UNLESS(stageInfo.Meta.IsSysView());
-
-    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
-
-    const auto& holderFactory = TxAlloc->HolderFactory;
-    const auto& typeEnv = TxAlloc->TypeEnv;
-    const auto& tableInfo = stageInfo.Meta.TableConstInfo;
-    const auto& keyTypes = tableInfo->KeyColumnTypes;
-
-    for (auto& op : stage.GetTableOps()) {
-        Y_DEBUG_ABORT_UNLESS(stageInfo.Meta.TablePath == op.GetTable().GetPath());
-
-        auto& task = AddTask(stageInfo, TTaskType::SYSVIEW_COMPUTE);
-        TShardKeyRanges keyRanges;
-
-        switch (op.GetTypeCase()) {
-            case NKqpProto::TKqpPhyTableOperation::kReadRange:
-                stageInfo.Meta.SkipNullKeys.assign(
-                    op.GetReadRange().GetSkipNullKeys().begin(),
-                    op.GetReadRange().GetSkipNullKeys().end()
-                );
-                keyRanges.Add(MakeKeyRange(
-                    keyTypes, op.GetReadRange().GetKeyRange(),
-                    stageInfo, holderFactory, typeEnv)
-                );
-                break;
-            case NKqpProto::TKqpPhyTableOperation::kReadRanges:
-                keyRanges.CopyFrom(FillReadRanges(keyTypes, op.GetReadRanges(), stageInfo, typeEnv));
-                break;
-            default:
-                YQL_ENSURE(false, "Unexpected table scan operation: " << (ui32) op.GetTypeCase());
-        }
-
-        TTaskMeta::TShardReadInfo readInfo = {
-            .Ranges = std::move(keyRanges),
-            .Columns = BuildKqpColumns(op, tableInfo),
-        };
-
-        auto readSettings = ExtractReadSettings(op, stageInfo, holderFactory, typeEnv);
-        task.Meta.Reads.ConstructInPlace();
-        task.Meta.Reads->emplace_back(std::move(readInfo));
-        task.Meta.ReadInfo.SetSorting(readSettings.GetSorting());
-        task.Meta.Type = TTaskMeta::TTaskType::Compute;
-
-        LOG_D("Stage " << stageInfo.Id << " create sysview scan task: " << task.Id);
     }
 }
 
@@ -1941,108 +1670,8 @@ std::pair<ui32, TKqpTasksGraph::TTaskType::ECreateReason> TKqpTasksGraph::GetMax
     return {result, taskReason};
 }
 
-bool TKqpTasksGraph::BuildComputeTasks(TStageInfo& stageInfo, const ui32 nodesCount) {
-    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
-
-    TTaskType::ECreateReason tasksReason = TTaskType::MINIMUM_COMPUTE;
-    bool unknownAffectedShardCount = false;
-    ui32 partitionsCount = 1;
-    ui32 inputTasks = 0;
-    bool isShuffle = false;
-    bool forceMapTasks = false;
-    ui32 mapConnectionCount = 0;
-
-    for (ui32 inputIndex = 0; inputIndex < stage.InputsSize(); ++inputIndex) {
-        const auto& input = stage.GetInputs(inputIndex);
-
-        // Current assumptions:
-        // 1. All stage's inputs, except 1st one, must be a `Broadcast` or `UnionAll`
-        // 2. Stages where 1st input is `Broadcast` are not partitioned.
-        if (inputIndex > 0) {
-            switch (input.GetTypeCase()) {
-                case NKqpProto::TKqpPhyConnection::kBroadcast:
-                case NKqpProto::TKqpPhyConnection::kHashShuffle:
-                case NKqpProto::TKqpPhyConnection::kUnionAll:
-                case NKqpProto::TKqpPhyConnection::kMerge:
-                case NKqpProto::TKqpPhyConnection::kStreamLookup:
-                case NKqpProto::TKqpPhyConnection::kMap:
-                case NKqpProto::TKqpPhyConnection::kParallelUnionAll:
-                case NKqpProto::TKqpPhyConnection::kVectorResolve:
-                case NKqpProto::TKqpPhyConnection::kDqSourceStreamLookup:
-                    break;
-                default:
-                    YQL_ENSURE(false, "Unexpected connection type: " << (ui32)input.GetTypeCase() << Endl);
-                    // TODO: << this->DebugString());
-            }
-        }
-
-        auto& originStageInfo = GetStageInfo(NYql::NDq::TStageId(stageInfo.Id.TxId, input.GetStageIndex()));
-
-        switch (input.GetTypeCase()) {
-            case NKqpProto::TKqpPhyConnection::kHashShuffle: {
-                inputTasks += originStageInfo.Tasks.size();
-                isShuffle = true;
-                break;
-            }
-            case NKqpProto::TKqpPhyConnection::kStreamLookup: {
-                tasksReason = TTaskType::PREV_STAGE_COMPUTE;
-                partitionsCount = originStageInfo.Tasks.size();
-                unknownAffectedShardCount = true;
-                break;
-            }
-            case NKqpProto::TKqpPhyConnection::kMap: {
-                tasksReason = TTaskType::PREV_STAGE_COMPUTE;
-                partitionsCount = originStageInfo.Tasks.size();
-                forceMapTasks = true;
-                ++mapConnectionCount;
-                break;
-            }
-            case NKqpProto::TKqpPhyConnection::kParallelUnionAll: {
-                partitionsCount = std::max<ui64>(partitionsCount, originStageInfo.Tasks.size());
-                break;
-            }
-            case NKqpProto::TKqpPhyConnection::kVectorResolve: {
-                tasksReason = TTaskType::PREV_STAGE_COMPUTE;
-                partitionsCount = originStageInfo.Tasks.size();
-                unknownAffectedShardCount = true;
-                break;
-            }
-            case NKqpProto::TKqpPhyConnection::kSequencer: {
-                tasksReason = TTaskType::PREV_STAGE_COMPUTE;
-                partitionsCount = originStageInfo.Tasks.size();
-                break;
-            }
-            default:
-                break;
-        }
-    }
-
-    Y_ENSURE(mapConnectionCount <= 1, "Only a single map connection is allowed");
-
-    if (isShuffle && !forceMapTasks) {
-        if (stage.GetTaskCount()) {
-            tasksReason = TTaskType::FORCED;
-            partitionsCount = stage.GetTaskCount();
-        } else {
-            auto [newPartitionCount, minTasksReason] = GetMaxTasksAggregation(stageInfo, inputTasks, nodesCount);
-            if (newPartitionCount > partitionsCount) {
-                tasksReason = minTasksReason;
-                partitionsCount = newPartitionCount;
-            }
-        }
-    }
-
-    for (ui32 i = 0; i < partitionsCount; ++i) {
-        auto& task = AddTask(stageInfo, tasksReason);
-        task.Meta.Type = TTaskMeta::TTaskType::Compute;
-        LOG_D("Stage " << stageInfo.Id << " create compute task: " << task.Id);
-    }
-
-    return unknownAffectedShardCount;
-}
-
 std::pair<ui32, TKqpTasksGraph::TTaskType::ECreateReason> TKqpTasksGraph::GetScanTasksPerNode(TStageInfo& stageInfo, const bool isOlapScan, const ui64 /* nodeId */, bool enableShuffleElimination) const {
-    TTaskType::ECreateReason taskReason;
+    TTaskType::ECreateReason taskReason = TTaskType::UNKNOWN;
     const auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
     if (const auto taskCount = stage.GetTaskCount()) {
         taskReason = TTaskType::FORCED;
@@ -2089,207 +1718,7 @@ std::pair<ui32, TKqpTasksGraph::TTaskType::ECreateReason> TKqpTasksGraph::GetSca
     return {result, taskReason};
 }
 
-void TKqpTasksGraph::BuildScanTasksFromShards(TStageInfo& stageInfo, bool enableShuffleElimination, TQueryExecutionStats* stats) {
-    THashMap<ui64, std::vector<ui64>> nodeTasks;
-    THashMap<ui64 /* nodeId */, std::vector<TShardInfoWithId>> nodeShards;
-    THashMap<ui64, ui64> assignedShardsCount;
-    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
-
-    auto& columnShardHashV1Params = stageInfo.Meta.ColumnShardHashV1Params;
-    bool shuffleEliminated = enableShuffleElimination && stage.GetIsShuffleEliminated();
-    if (shuffleEliminated && stageInfo.Meta.ColumnTableInfoPtr) {
-        const auto& tableDesc = stageInfo.Meta.ColumnTableInfoPtr->Description;
-        columnShardHashV1Params.SourceShardCount = tableDesc.GetColumnShardCount();
-        columnShardHashV1Params.SourceTableKeyColumnTypes = std::make_shared<TVector<NScheme::TTypeInfo>>();
-        for (const auto& column: tableDesc.GetSharding().GetHashSharding().GetColumns()) {
-            Y_ENSURE(stageInfo.Meta.TableConstInfo->Columns.contains(column), TStringBuilder{} << "Table doesn't have column: " << column);
-            auto columnType = stageInfo.Meta.TableConstInfo->Columns.at(column).Type;
-            columnShardHashV1Params.SourceTableKeyColumnTypes->push_back(columnType);
-        }
-    }
-
-    const auto& tableInfo = stageInfo.Meta.TableConstInfo;
-    const auto& keyTypes = tableInfo->KeyColumnTypes;
-    for (int i = 0, s = stage.TableOpsSize(); i < s; ++i) {
-        const auto& op = stage.GetTableOps(i);
-        Y_DEBUG_ABORT_UNLESS(stageInfo.Meta.TablePath == op.GetTable().GetPath());
-
-        auto columns = BuildKqpColumns(op, tableInfo);
-        auto& partitions = stageInfo.Meta.PrunedPartitions.at(i);
-        const bool isOlapScan = (op.GetTypeCase() == NKqpProto::TKqpPhyTableOperation::kReadOlapRange);
-        auto readSettings = ExtractReadSettings(op, stageInfo, TxAlloc->HolderFactory, TxAlloc->TypeEnv);
-
-        if (op.GetTypeCase() == NKqpProto::TKqpPhyTableOperation::kReadRange) {
-            stageInfo.Meta.SkipNullKeys.assign(op.GetReadRange().GetSkipNullKeys().begin(), op.GetReadRange().GetSkipNullKeys().end());
-            YQL_ENSURE(!readSettings.IsReverse(), "Not supported for scan queries");
-        }
-
-        YQL_ENSURE(GetMeta().ShardsResolved);
-
-        for (auto&& [shardId, shardInfo]: partitions) {
-            const ui64 nodeId = GetMeta().ShardIdToNodeId.at(shardId);
-            nodeShards[nodeId].emplace_back(TShardInfoWithId(shardId, std::move(shardInfo)));
-        }
-
-        if (stats) {
-            for (const auto& [nodeId, shardsInfo] : nodeShards) {
-                stats->AddNodeShardsCount(stageInfo.Id.StageId, nodeId, shardsInfo.size());
-            }
-        }
-
-        if (!AppData()->FeatureFlags.GetEnableSeparationComputeActorsFromRead() && !shuffleEliminated || (!isOlapScan && readSettings.IsSorted())) {
-            THashMap<ui64 /* nodeId */, ui64 /* tasks count */> olapAndSortedTasksCount;
-
-            auto AssignScanTaskToShard = [&](const ui64 shardId, const bool sorted) -> TTask& {
-                ui64 nodeId = GetMeta().ShardIdToNodeId.at(shardId);
-                if (stageInfo.Meta.IsOlap() && sorted) {
-                    auto& task = AddTask(stageInfo, TTaskType::OLAP_SORT_SCAN);
-                    task.Meta.NodeId = nodeId;
-                    task.Meta.ScanTask = true;
-                    task.Meta.Type = TTaskMeta::TTaskType::Scan;
-                    ++olapAndSortedTasksCount[nodeId];
-                    return task;
-                }
-
-                auto& tasks = nodeTasks[nodeId];
-                auto& cnt = assignedShardsCount[nodeId];
-                const auto [maxScansPerNode, maxTasksReason] = GetScanTasksPerNode(stageInfo, isOlapScan, nodeId);
-                if (cnt < maxScansPerNode) {
-                    auto& task = AddTask(stageInfo, TTaskType::DEFAULT_SHARD_SCAN);
-                    task.Meta.NodeId = nodeId;
-                    task.Meta.ScanTask = true;
-                    task.Meta.Type = TTaskMeta::TTaskType::Scan;
-                    tasks.push_back(task.Id);
-                    ++cnt;
-
-                    if (cnt == maxScansPerNode) {
-                        for (const auto& taskId : tasks) {
-                            GetTask(taskId).Reason = maxTasksReason;
-                        }
-                    }
-
-                    return task;
-                } else {
-                    ui64 taskIdx = cnt % maxScansPerNode;
-                    ++cnt;
-                    return GetTask(tasks[taskIdx]);
-                }
-            };
-
-            for (auto& [_, shardsInfo] : nodeShards) {
-                for (auto& shardInfo : shardsInfo) {
-                    auto& task = AssignScanTaskToShard(shardInfo.ShardId, readSettings.IsSorted());
-                    MergeReadInfoToTaskMeta(task.Meta, shardInfo.ShardId, shardInfo.KeyReadRanges, readSettings,
-                        columns, op, /*isPersistentScan*/ true);
-                }
-            }
-
-            for (const auto& [nodeId, tasksId] : nodeTasks) {
-                for (const auto& taskIdx : tasksId) {
-                    auto& task = GetTask(taskIdx);
-                    task.Meta.SetEnableShardsSequentialScan(readSettings.IsSorted());
-                    PrepareScanMetaForUsage(task.Meta, keyTypes);
-                }
-            }
-        } else if (shuffleEliminated /* save partitioning for shuffle elimination */) {
-            std::size_t stageInternalTaskId = 0;
-            columnShardHashV1Params.TaskIndexByHash = std::make_shared<TVector<ui64>>();
-            columnShardHashV1Params.TaskIndexByHash->resize(columnShardHashV1Params.SourceShardCount);
-
-            for (auto&& [nodeId, shardsInfo] : nodeShards) {
-                auto tasksReason = TTaskType::SHUFFLE_ELIMINATE_SCAN;
-                std::size_t tasksPerNode = shardsInfo.size();
-                auto [maxTasksPerNode, maxTasksReason] = GetScanTasksPerNode(stageInfo, isOlapScan, nodeId, true);
-
-                if (maxTasksPerNode < tasksPerNode) {
-                    tasksReason = maxTasksReason;
-                    tasksPerNode = maxTasksPerNode;
-                }
-
-                std::vector<TTaskMeta> metas(tasksPerNode, TTaskMeta());
-                {
-                    for (std::size_t i = 0; i < shardsInfo.size(); ++i) {
-                        auto&& shardInfo = shardsInfo[i];
-                        MergeReadInfoToTaskMeta(
-                            metas[i % tasksPerNode],
-                            shardInfo.ShardId,
-                            shardInfo.KeyReadRanges,
-                            readSettings,
-                            columns, op,
-                            /*isPersistentScan*/ true
-                        );
-                    }
-
-                    for (auto& meta: metas) {
-                        PrepareScanMetaForUsage(meta, keyTypes);
-                        LOG_D( "Stage " << stageInfo.Id << " create scan task meta for node: " << nodeId
-                            << ", meta: " << meta.ToString(keyTypes, *AppData()->TypeRegistry));
-                    }
-                }
-
-                // in runtime we calc hash, which will be in [0; shardcount]
-                // so we merge to mappings : hash -> shardID and shardID -> channelID for runtime
-                THashMap<ui64, ui64> hashByShardId;
-                Y_ENSURE(stageInfo.Meta.ColumnTableInfoPtr != nullptr, "ColumnTableInfoPtr is nullptr, maybe information about shards haven't beed delivered yet.");
-                const auto& tableDesc = stageInfo.Meta.ColumnTableInfoPtr->Description;
-                const auto& sharding = tableDesc.GetSharding();
-                for (std::size_t i = 0; i < sharding.ColumnShardsSize(); ++i) {
-                    hashByShardId.insert({sharding.GetColumnShards(i), i});
-                }
-
-                for (ui32 t = 0; t < tasksPerNode; ++t, ++stageInternalTaskId) {
-                    auto& task = AddTask(stageInfo, tasksReason);
-                    task.Meta = metas[t];
-                    task.Meta.SetEnableShardsSequentialScan(false);
-                    task.Meta.NodeId = nodeId;
-                    task.Meta.ScanTask = true;
-                    task.Meta.Type = TTaskMeta::TTaskType::Scan;
-                    task.SetMetaId(t);
-
-                    for (const auto& readInfo: *task.Meta.Reads) {
-                        Y_ENSURE(hashByShardId.contains(readInfo.ShardId));
-                        (*columnShardHashV1Params.TaskIndexByHash)[hashByShardId[readInfo.ShardId]] = stageInternalTaskId;
-                    }
-                }
-            }
-
-            LOG_D( "Stage with scan " << "[" << stageInfo.Id.TxId << ":" << stageInfo.Id.StageId << "]"
-                << " has keys: " << columnShardHashV1Params.KeyTypesToString() << " and task count: " << stageInternalTaskId;
-            );
-        } else {
-            ui32 metaId = 0;
-            for (auto&& [nodeId, shardsInfo] : nodeShards) {
-                const ui32 metaGlueingId = ++metaId;
-                TTaskMeta meta;
-                {
-                    for (auto&& shardInfo : shardsInfo) {
-                        MergeReadInfoToTaskMeta(meta, shardInfo.ShardId, shardInfo.KeyReadRanges, readSettings,
-                            columns, op, /*isPersistentScan*/ true);
-                    }
-                    PrepareScanMetaForUsage(meta, keyTypes);
-                    LOG_D( "Stage " << stageInfo.Id << " create scan task meta for node: " << nodeId
-                        << ", meta: " << meta.ToString(keyTypes, *AppData()->TypeRegistry));
-                }
-
-                const auto [maxTasksPerNode, tasksReason] = GetScanTasksPerNode(stageInfo, isOlapScan, nodeId);
-
-                for (ui32 t = 0; t < maxTasksPerNode; ++t) {
-                    auto& task = AddTask(stageInfo, tasksReason);
-                    task.Meta = meta;
-                    task.Meta.SetEnableShardsSequentialScan(false);
-                    task.Meta.NodeId = nodeId;
-                    task.Meta.ScanTask = true;
-                    task.Meta.Type = TTaskMeta::TTaskType::Scan;
-                    task.SetMetaId(metaGlueingId);
-                }
-            }
-        }
-    }
-
-    LOG_D("Stage " << stageInfo.Id << " will be executed on " << nodeTasks.size() << " nodes.");
-}
-
-void TKqpTasksGraph::BuildReadTasksFromSource(TStageInfo& stageInfo, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot, ui32 scheduledTaskCount) {
+void TKqpTasksGraph::RestoreReadTasksFromSource(TStageInfo& stageInfo, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot, ui32 scheduledTaskCount) {
     const auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
 
     YQL_ENSURE(stage.GetSources(0).HasExternalSource());
@@ -2298,26 +1727,21 @@ void TKqpTasksGraph::BuildReadTasksFromSource(TStageInfo& stageInfo, const TVect
     const auto& stageSource = stage.GetSources(0);
     const auto& externalSource = stageSource.GetExternalSource();
 
-    auto tasksReason = TTaskType::DEFAULT_SOURCE_READ;
     ui32 taskCount = externalSource.GetPartitionedTaskParams().size();
 
-    auto tasksReasonHint = TTaskType::FORCED;
     auto taskCountHint = stage.GetTaskCount();
 
     if (taskCountHint == 0) {
-        tasksReasonHint = TTaskType::SCHEDULED_SOURCE_READ;
         taskCountHint = scheduledTaskCount;
     }
 
     if (taskCountHint) {
         if (taskCount > taskCountHint) {
-            tasksReason = tasksReasonHint;
             taskCount = taskCountHint;
         }
     } else if (!resourceSnapshot.empty()) {
         ui32 maxTaskCount = resourceSnapshot.size() * 2;
         if (taskCount > maxTaskCount) {
-            tasksReason = TTaskType::SNAPSHOT_SOURCE_READ;
             taskCount = maxTaskCount;
         }
     }
@@ -2336,264 +1760,9 @@ void TKqpTasksGraph::BuildReadTasksFromSource(TStageInfo& stageInfo, const TVect
         }
     }
 
-    if (GetMeta().IsRestored) {
-        for (const auto taskId : stageInfo.Tasks) {
-            FillReadTaskFromSource(GetTask(taskId), sourceName, structuredToken, resourceSnapshot, nodeOffset++);
-        }
-        return;
+    for (const auto taskId : stageInfo.Tasks) {
+        FillReadTaskFromSource(GetTask(taskId), sourceName, structuredToken, resourceSnapshot, nodeOffset++);
     }
-
-    TVector<ui64> tasksIds;
-    tasksIds.reserve(taskCount);
-
-    // generate all tasks
-    for (ui32 i = 0; i < taskCount; i++) {
-        auto& task = AddTask(stageInfo, tasksReason);
-
-        if (!externalSource.GetEmbedded()) {
-            auto& input = task.Inputs[stageSource.GetInputIndex()];
-            input.ConnectionInfo = NYql::NDq::TSourceInput{};
-            input.SourceSettings = externalSource.GetSettings();
-            input.SourceType = externalSource.GetType();
-            if (externalSource.HasWatermarksSettings()) {
-                const auto& watermarksSettings = externalSource.GetWatermarksSettings();
-                input.WatermarksMode = NYql::NDqProto::EWatermarksMode::WATERMARKS_MODE_DEFAULT;
-                if (watermarksSettings.HasIdleTimeoutUs()) {
-                    input.WatermarksIdleTimeoutUs = watermarksSettings.GetIdleTimeoutUs();
-                }
-            }
-        }
-
-        FillReadTaskFromSource(task, sourceName, structuredToken, resourceSnapshot, nodeOffset++);
-
-        TString queryPath = "default";
-        if (GetMeta().UserRequestContext && GetMeta().UserRequestContext->StreamingQueryPath) {
-            queryPath = GetMeta().UserRequestContext->StreamingQueryPath;
-        }
-        task.Meta.TaskParams.emplace("query_path", queryPath);
-
-        tasksIds.push_back(task.Id);
-    }
-
-    // distribute read ranges between them
-    ui32 currentTaskIndex = 0;
-    for (const TString& partitionParam : externalSource.GetPartitionedTaskParams()) {
-        GetTask(tasksIds[currentTaskIndex]).Meta.ReadRanges.push_back(partitionParam);
-        if (++currentTaskIndex >= tasksIds.size()) {
-            currentTaskIndex = 0;
-        }
-    }
-}
-
-void TKqpTasksGraph::BuildFullTextScanTasksFromSource(TStageInfo& stageInfo, TQueryExecutionStats* stats) {
-    YQL_ENSURE(stageInfo.Meta.GetStage(stageInfo.Id).GetSources(0).GetTypeCase() == NKqpProto::TKqpSource::kFullTextSource);
-    Y_UNUSED(stats);
-
-    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
-    const auto& source = stageInfo.Meta.GetStage(stageInfo.Id).GetSources(0);
-    const auto& fullTextSource = source.GetFullTextSource();
-
-    YQL_ENSURE(fullTextSource.GetIndex());
-
-    auto& task = AddTask(stageInfo, TTaskType::DEFAULT_SOURCE_READ);
-    task.Meta.Type = TTaskMeta::TTaskType::Scan;
-    task.Meta.NodeId = GetMeta().ExecuterId.NodeId();
-    const auto& stageSource = stage.GetSources(0);
-    auto& input = task.Inputs.at(stageSource.GetInputIndex());
-    input.SourceType = NYql::KqpFullTextSourceName;
-    input.ConnectionInfo = NYql::NDq::TSourceInput{};
-
-    input.Meta.FullTextSourceSettings = GetMeta().Allocate<NKikimrKqp::TKqpFullTextSourceSettings>();
-    NKikimrKqp::TKqpFullTextSourceSettings* settings = input.Meta.FullTextSourceSettings;
-
-    settings->SetIndex(fullTextSource.GetIndex());
-    settings->SetDatabase(GetMeta().Database);
-    settings->MutableTable()->CopyFrom(fullTextSource.GetTable());
-    settings->SetIndexType(fullTextSource.GetIndexType());
-    settings->MutableIndexDescription()->CopyFrom(fullTextSource.GetIndexDescription());
-
-    auto guard = TxAlloc->TypeEnv.BindAllocator();
-    {
-
-        TStringBuilder queryBuilder;
-        for (const auto& query : fullTextSource.GetQuerySettings().GetQueryValue()) {
-            auto value = ExtractPhyValue(
-                stageInfo, query,
-                TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod());
-            queryBuilder << TString(value.AsStringRef());
-        }
-
-        settings->MutableQuerySettings()->SetQuery(TString(queryBuilder));
-    }
-
-    if (fullTextSource.HasTakeLimit())
-    {
-        auto value = ExtractPhyValue(
-            stageInfo, fullTextSource.GetTakeLimit(),
-            TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod());
-        if (value.HasValue()) {
-            settings->SetLimit(value.Get<ui64>());
-        }
-    }
-
-    if (fullTextSource.HasBFactor()) {
-        auto value = ExtractPhyValue(
-            stageInfo, fullTextSource.GetBFactor(),
-            TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod());
-
-        if (value.HasValue()) {
-            settings->SetBFactor(value.Get<double>());
-        }
-    }
-
-    if (fullTextSource.HasDefaultOperator()) {
-        auto value = ExtractPhyValue(
-            stageInfo, fullTextSource.GetDefaultOperator(),
-            TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod());
-        if (value.HasValue()) {
-            settings->SetDefaultOperator(TString(value.AsStringRef()));
-        }
-    }
-
-    if (fullTextSource.HasMinimumShouldMatch()) {
-        auto value = ExtractPhyValue(
-            stageInfo, fullTextSource.GetMinimumShouldMatch(),
-            TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod());
-        if (value.HasValue()) {
-            settings->SetMinimumShouldMatch(TString(value.AsStringRef()));
-        }
-    }
-
-    if (fullTextSource.HasK1Factor()) {
-        auto value = ExtractPhyValue(
-            stageInfo, fullTextSource.GetK1Factor(),
-            TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod());
-        if (value.HasValue()) {
-            settings->SetK1Factor(value.Get<double>());
-        }
-    }
-
-    settings->MutableQuerySettings()->MutableColumns()->CopyFrom(fullTextSource.GetQuerySettings().GetColumns());
-
-    for(const auto& indexTable : fullTextSource.GetIndexTables()) {
-        auto* indexTableProto = settings->AddIndexTables();
-        indexTableProto->MutableTable()->CopyFrom(indexTable.GetTable());
-        indexTableProto->MutableKeyColumns()->CopyFrom(indexTable.GetKeyColumns());
-        indexTableProto->MutableColumns()->CopyFrom(indexTable.GetColumns());
-    }
-
-    settings->MutableKeyColumns()->CopyFrom(fullTextSource.GetKeyColumns());
-    settings->MutableColumns()->CopyFrom(fullTextSource.GetColumns());
-    if (GetMeta().Snapshot.IsValid()) {
-        settings->MutableSnapshot()->SetStep(GetMeta().Snapshot.Step);
-        settings->MutableSnapshot()->SetTxId(GetMeta().Snapshot.TxId);
-    }
-}
-
-void TKqpTasksGraph::BuildSysViewTasksFromSource(TStageInfo& stageInfo) {
-    YQL_ENSURE(stageInfo.Meta.GetStage(stageInfo.Id).GetSources(0).GetTypeCase() == NKqpProto::TKqpSource::kSysViewSource);
-
-    auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
-    const auto& source = stage.GetSources(0);
-    const auto& sysViewSource = source.GetSysViewSource();
-
-    auto& task = AddTask(stageInfo, TTaskType::DEFAULT_SOURCE_READ);
-    task.Meta.Type = TTaskMeta::TTaskType::Compute;
-    task.Meta.NodeId = GetMeta().ExecuterId.NodeId();
-
-    const auto& stageSource = stage.GetSources(0);
-    auto& input = task.Inputs.at(stageSource.GetInputIndex());
-    input.SourceType = NYql::KqpSysViewSourceName;
-    input.ConnectionInfo = NYql::NDq::TSourceInput{};
-
-    input.Meta.SysViewSourceSettings = GetMeta().Allocate<NKikimrKqp::TKqpSysViewSourceSettings>();
-    NKikimrKqp::TKqpSysViewSourceSettings* settings = input.Meta.SysViewSourceSettings;
-
-    settings->SetDatabase(GetMeta().Database);
-    settings->MutableTable()->CopyFrom(sysViewSource.GetTable());
-    settings->SetTablePath(sysViewSource.GetTable().GetPath());
-    settings->SetSysViewInfo(sysViewSource.GetTable().GetSysView());
-    settings->SetReverse(sysViewSource.GetReverse());
-
-    // Fill SysViewDescription from table metadata
-    const auto& tableInfo = stageInfo.Meta.TableConstInfo;
-    if (tableInfo && tableInfo->SysViewInfo) {
-        *settings->MutableSysViewDescription() = *tableInfo->SysViewInfo;
-    }
-
-    // Fill columns: convert TKqpPhyColumnId to TKqpColumnMetadataProto using table metadata
-    for (const auto& phyCol : sysViewSource.GetColumns()) {
-        auto* col = settings->AddColumns();
-        col->SetId(phyCol.GetId());
-        col->SetName(phyCol.GetName());
-        if (tableInfo) {
-            auto it = tableInfo->Columns.find(phyCol.GetName());
-            if (it != tableInfo->Columns.end()) {
-                col->SetTypeId(it->second.Type.GetTypeId());
-                if (NScheme::NTypeIds::IsParametrizedType(it->second.Type.GetTypeId())) {
-                    NScheme::ProtoFromTypeInfo(it->second.Type, {}, *col->MutableTypeInfo());
-                }
-            }
-        }
-    }
-
-    // Fill key columns from table info
-    if (tableInfo) {
-        for (size_t i = 0; i < tableInfo->KeyColumns.size(); ++i) {
-            auto* kc = settings->AddKeyColumns();
-            kc->SetName(tableInfo->KeyColumns[i]);
-            if (i < tableInfo->KeyColumnTypes.size()) {
-                kc->SetTypeId(tableInfo->KeyColumnTypes[i].GetTypeId());
-            }
-        }
-    }
-
-    // Fill key ranges from the source proto
-    const auto& holderFactory = TxAlloc->HolderFactory;
-    const auto& typeEnv = TxAlloc->TypeEnv;
-    const auto& keyTypes = tableInfo ? tableInfo->KeyColumnTypes : TVector<NScheme::TTypeInfo>();
-
-    auto guard = TxAlloc->TypeEnv.BindAllocator();
-
-    switch (sysViewSource.GetRangesExprCase()) {
-        case NKqpProto::TKqpSysViewSource::kKeyRange: {
-            auto range = MakeKeyRange(keyTypes, sysViewSource.GetKeyRange(), stageInfo, holderFactory, typeEnv);
-            range.Serialize(*settings->AddKeyRanges());
-            break;
-        }
-        case NKqpProto::TKqpSysViewSource::kRanges: {
-            auto ranges = FillRangesFromParameter(keyTypes, sysViewSource.GetRanges(), stageInfo, typeEnv);
-            for (auto& pointOrRange : ranges) {
-                if (auto* range = std::get_if<TSerializedTableRange>(&pointOrRange)) {
-                    range->Serialize(*settings->AddKeyRanges());
-                } else {
-                    // Convert point to an inclusive range [point, point]
-                    auto& point = std::get<TSerializedCellVec>(pointOrRange);
-                    TSerializedTableRange rangeFromPoint(point.GetCells(), true, point.GetCells(), true);
-                    rangeFromPoint.Point = true;
-                    rangeFromPoint.Serialize(*settings->AddKeyRanges());
-                }
-            }
-            break;
-        }
-        default: {
-            // Full scan — construct a range from [NULL..NULL] (beginning) to [] (end), both inclusive.
-            // Scan actors expect From to have at least one NULL cell for full scans;
-            // a default-constructed TSerializedTableRange has empty From cells which some
-            // scan actors interpret as +inf.
-            TVector<TCell> fromCells(keyTypes.size()); // NULL cells = "from the very beginning"
-            TSerializedTableRange fullRange(fromCells, true, TConstArrayRef<TCell>(), true);
-            fullRange.Serialize(*settings->AddKeyRanges());
-            break;
-        }
-    }
-
-    // Pass user token for access control in scan actors (e.g., Auth* sys views)
-    if (UserToken) {
-        settings->SetUserToken(UserToken->SerializeAsString());
-    }
-
-    LOG_D("Stage " << stageInfo.Id << " create sys view source task: " << task.Id);
 }
 
 void TKqpTasksGraph::FillScanTaskLockTxId(NKikimrTxDataShard::TKqpReadRangesSourceSettings& settings) {
@@ -2605,272 +1774,6 @@ void TKqpTasksGraph::FillScanTaskLockTxId(NKikimrTxDataShard::TKqpReadRangesSour
         GetMeta().QuerySpanId, settings.GetTable().GetTablePath());
     if (effectiveSpanId) {
         settings.SetQuerySpanId(effectiveSpanId);
-    }
-}
-
-TMaybe<size_t> TKqpTasksGraph::BuildScanTasksFromSource(TStageInfo& stageInfo, bool limitTasksPerNode, TQueryExecutionStats* stats) {
-    THashMap<ui64, std::vector<ui64>> nodeTasks;
-    THashMap<ui64, ui64> assignedShardsCount;
-
-    const auto& stage = stageInfo.Meta.GetStage(stageInfo.Id);
-
-    bool singlePartitionedStage = stage.GetIsSinglePartition();
-
-    // TODO: describe in comments the exact expected stage state.
-    YQL_ENSURE(stage.GetSources(0).HasReadRangesSource());
-    YQL_ENSURE(stage.GetSources(0).GetInputIndex() == 0 && stage.SourcesSize() == 1);
-    for (const auto& input : stage.GetInputs()) {
-        YQL_ENSURE(input.HasBroadcast());
-    }
-
-    const auto& source = stage.GetSources(0).GetReadRangesSource();
-    const auto& tableInfo = stageInfo.Meta.TableConstInfo;
-    const auto& keyTypes = tableInfo->KeyColumnTypes;
-
-    // TODO: what is the difference with `stageInfo.Meta.IsOlap()`?
-    YQL_ENSURE(tableInfo->TableKind != NKikimr::NKqp::ETableKind::Olap);
-
-    auto columns = BuildKqpColumns(source, tableInfo);
-
-    const auto& snapshot = GetMeta().Snapshot;
-
-    TVector<ui64> createdTasksIds;
-    auto createNewTask = [&](
-            TMaybe<ui64> nodeId,
-            ui64 taskShardId,
-            bool hintShardId,
-            TMaybe<ui64> maxInFlightShards,
-            TTaskType::ECreateReason taskReason) -> TTask&
-    {
-        auto& task = AddTask(stageInfo, taskReason);
-        task.Meta.Type = TTaskMeta::TTaskType::Scan; // TODO: can this scan task not have NodeId?
-        if (nodeId) {
-            task.Meta.NodeId = *nodeId;
-        }
-
-        if (!nodeId || !GetMeta().ShardsResolved) {
-            YQL_ENSURE(!GetMeta().ShardsResolved);
-            task.Meta.ShardId = taskShardId;
-        }
-
-        const auto& stageSource = stage.GetSources(0);
-        auto& input = task.Inputs.at(stageSource.GetInputIndex());
-        input.SourceType = NYql::KqpReadRangesSourceName;
-        input.ConnectionInfo = NYql::NDq::TSourceInput{};
-
-        // allocating source settings
-
-        input.Meta.SourceSettings = GetMeta().Allocate<NKikimrTxDataShard::TKqpReadRangesSourceSettings>();
-        NKikimrTxDataShard::TKqpReadRangesSourceSettings* settings = input.Meta.SourceSettings;
-        settings->SetDatabase(GetMeta().Database);
-        FillTableMeta(stageInfo, settings->MutableTable());
-
-        settings->SetIsTableImmutable(source.GetIsTableImmutable());
-        settings->SetIsolationLevel(GetMeta().RequestIsolationLevel);
-
-        for (auto& keyColumn : keyTypes) {
-            auto columnType = NScheme::ProtoColumnTypeFromTypeInfoMod(keyColumn, "");
-            *settings->AddKeyColumnTypeInfos() = columnType.TypeInfo ?
-                *columnType.TypeInfo :
-                NKikimrProto::TTypeInfo();
-            settings->AddKeyColumnTypes(static_cast<ui32>(keyColumn.GetTypeId()));
-        }
-
-        for (auto& column : columns) {
-            auto* protoColumn = settings->AddColumns();
-            protoColumn->SetId(column.Id);
-            auto columnType = NScheme::ProtoColumnTypeFromTypeInfoMod(column.Type, column.TypeMod);
-            protoColumn->SetType(columnType.TypeId);
-            protoColumn->SetNotNull(column.NotNull);
-            protoColumn->SetIsPrimary(column.IsPrimary);
-            if (columnType.TypeInfo) {
-                *protoColumn->MutableTypeInfo() = *columnType.TypeInfo;
-            }
-            protoColumn->SetName(column.Name);
-        }
-
-        if (GetMeta().CheckDuplicateRows) {
-            for (auto& colName : tableInfo->KeyColumns) {
-                const auto& tableColumn = tableInfo->Columns.at(colName);
-                auto* protoColumn = settings->AddDuplicateCheckColumns();
-                protoColumn->SetId(tableColumn.Id);
-                auto columnType = NScheme::ProtoColumnTypeFromTypeInfoMod(tableColumn.Type, tableColumn.TypeMod);
-                protoColumn->SetType(columnType.TypeId);
-                if (columnType.TypeInfo) {
-                    *protoColumn->MutableTypeInfo() = *columnType.TypeInfo;
-                }
-                protoColumn->SetName(colName);
-            }
-        }
-
-        if (AppData()->FeatureFlags.GetEnableArrowFormatAtDatashard()) {
-            settings->SetDataFormat(NKikimrDataEvents::FORMAT_ARROW);
-        } else {
-            settings->SetDataFormat(NKikimrDataEvents::FORMAT_CELLVEC);
-        }
-
-        if (snapshot.IsValid()) {
-            settings->MutableSnapshot()->SetStep(snapshot.Step);
-            settings->MutableSnapshot()->SetTxId(snapshot.TxId);
-        }
-
-        if (GetMeta().RequestIsolationLevel == NKqpProto::ISOLATION_LEVEL_READ_UNCOMMITTED) {
-            settings->SetAllowInconsistentReads(true);
-        }
-
-        settings->SetReverse(source.GetReverse());
-        settings->SetSorted(source.GetSorted());
-
-        if (maxInFlightShards) {
-            settings->SetMaxInFlightShards(*maxInFlightShards);
-        }
-
-        if (hintShardId) {
-            settings->SetShardIdHint(taskShardId);
-        }
-
-        if (GetMeta().MaxBatchSize) {
-            settings->SetItemsLimit(*GetMeta().MaxBatchSize);
-            settings->SetIsBatch(true);
-        } else {
-            ui64 itemsLimit = ExtractPhyValue(stageInfo, source.GetItemsLimit(), TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod((ui32)0)).Get<ui64>();
-            settings->SetItemsLimit(itemsLimit);
-        }
-
-        if (source.HasVectorTopK()) {
-            const auto& in = source.GetVectorTopK();
-            auto& out = *settings->MutableVectorTopK();
-            out.SetColumn(in.GetColumn());
-            *out.MutableSettings() = in.GetSettings();
-            auto target = ExtractPhyValue(stageInfo, in.GetTargetVector(), TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod());
-            out.SetTargetVector(TString(target.AsStringRef()));
-            out.SetLimit((ui32)ExtractPhyValue(stageInfo, in.GetLimit(), TxAlloc->HolderFactory, TxAlloc->TypeEnv, NUdf::TUnboxedValuePod()).Get<ui64>());
-        }
-
-        FillScanTaskLockTxId(*settings);
-
-        if (GetMeta().LockMode) {
-            settings->SetLockMode(*GetMeta().LockMode);
-        }
-
-        createdTasksIds.push_back(task.Id);
-        return task;
-    };
-
-    THashMap<ui64, TVector<ui64>> nodeIdToTasks;
-    THashMap<ui64, TVector<TShardRangesWithShardId>> nodeIdToShardKeyRanges;
-
-    auto addPartition = [&](
-        ui64 taskShardId,
-        TMaybe<ui64> nodeId,
-        bool hintShardId,
-        const TShardInfo& shardInfo,
-        TMaybe<ui64> maxInFlightShards,
-        TTaskType::ECreateReason tasksReason)
-    {
-        YQL_ENSURE(!shardInfo.KeyWriteRanges);
-
-        if (!nodeId) {
-            const auto nodeIdPtr = GetMeta().ShardIdToNodeId.FindPtr(taskShardId);
-            nodeId = nodeIdPtr ? TMaybe<ui64>{*nodeIdPtr} : Nothing();
-        }
-
-        YQL_ENSURE(!GetMeta().ShardsResolved || nodeId);
-
-        if (hintShardId && stats) {
-            stats->AffectedShards.insert(taskShardId);
-        }
-
-        if (limitTasksPerNode && GetMeta().ShardsResolved) {
-            const auto [maxScanTasksPerNode, maxTasksReason] = GetScanTasksPerNode(stageInfo, /* isOlapScan */ false, *nodeId);
-            auto& nodeTasks = nodeIdToTasks[*nodeId];
-            if (nodeTasks.size() < maxScanTasksPerNode) {
-                const auto& task = createNewTask(nodeId, taskShardId, false, maxInFlightShards, tasksReason);
-                nodeTasks.push_back(task.Id);
-
-                if (nodeTasks.size() == maxScanTasksPerNode) {
-                    for (const auto& taskId : nodeTasks) {
-                        GetTask(taskId).Reason = maxTasksReason;
-                    }
-                }
-            }
-            nodeIdToShardKeyRanges[*nodeId].push_back(TShardRangesWithShardId{hintShardId ? TMaybe<ui64>(taskShardId) : Nothing(), &*shardInfo.KeyReadRanges});
-        } else {
-            auto& task = createNewTask(nodeId, taskShardId, hintShardId, maxInFlightShards, tasksReason);
-            const auto& stageSource = stage.GetSources(0);
-            auto& input = task.Inputs[stageSource.GetInputIndex()];
-            NKikimrTxDataShard::TKqpReadRangesSourceSettings* settings = input.Meta.SourceSettings;
-
-            shardInfo.KeyReadRanges->SerializeTo(settings);
-        }
-    };
-
-    auto fillRangesForTasks = [&]() {
-        for (const auto& [nodeId, shardsRanges] : nodeIdToShardKeyRanges) {
-            const auto& tasks = nodeIdToTasks.at(nodeId);
-
-            const auto rangesDistribution = DistributeShardsToTasks(shardsRanges, tasks.size(), keyTypes);
-            YQL_ENSURE(rangesDistribution.size() == tasks.size());
-
-            for (size_t taskIndex = 0; taskIndex < tasks.size(); ++taskIndex) {
-                const auto taskId = tasks[taskIndex];
-                auto& task = GetTask(taskId);
-                const auto& stageSource = stage.GetSources(0);
-                auto& input = task.Inputs[stageSource.GetInputIndex()];
-                NKikimrTxDataShard::TKqpReadRangesSourceSettings* settings = input.Meta.SourceSettings;
-
-                const auto& shardsRangesForTask = rangesDistribution[taskIndex];
-
-                if (shardsRangesForTask.size() == 1 && shardsRangesForTask[0].ShardId) {
-                    settings->SetShardIdHint(*shardsRangesForTask[0].ShardId);
-                }
-
-                bool hasRanges = false;
-                for (const auto& shardRanges : shardsRangesForTask) {
-                    hasRanges |= shardRanges.Ranges->HasRanges();
-                }
-
-                for (const auto& shardRanges : shardsRangesForTask) {
-                    shardRanges.Ranges->SerializeTo(settings, !hasRanges);
-                }
-            }
-        }
-    };
-
-    const auto& partitions = stageInfo.Meta.PrunedPartitions.at(0);
-
-    bool isSequentialInFlight = source.GetSequentialInFlightShards() > 0 && partitions.size() > source.GetSequentialInFlightShards();
-
-    if (!partitions.empty() && (isSequentialInFlight || singlePartitionedStage)) {
-        Y_ENSURE(stageInfo.Meta.VirtualPartition);
-
-        auto startShard = stageInfo.Meta.VirtualPartition->ShardId;
-        auto& shardInfo = *stageInfo.Meta.VirtualPartition;
-
-        if (stats) {
-            for (const auto& [shardId, _] : partitions) {
-                stats->AffectedShards.insert(shardId);
-            }
-        }
-
-        TMaybe<ui64> inFlightShards = Nothing();
-        if (isSequentialInFlight) {
-            inFlightShards = source.GetSequentialInFlightShards();
-        }
-
-        Y_ENSURE(shardInfo.KeyReadRanges); // TODO: redundant check - remove later.
-
-        const TMaybe<ui64> nodeId = singlePartitionedStage ? TMaybe<ui64>{GetMeta().ExecuterId.NodeId()} : Nothing();
-        addPartition(startShard, nodeId, false, shardInfo, inFlightShards, TTaskType::SINGLE_SOURCE_SCAN);
-        fillRangesForTasks();
-
-        return singlePartitionedStage ? TMaybe<size_t>(partitions.size()) : Nothing();
-    } else {
-        for (const auto& [shardId, shardInfo] : partitions) {
-            addPartition(shardId, {}, true, shardInfo, {}, TTaskType::DEFAULT_SOURCE_SCAN);
-        }
-        fillRangesForTasks();
-        return partitions.size();
     }
 }
 
@@ -2907,7 +1810,7 @@ void TKqpTasksGraph::BuildExternalSinks(const NKqpProto::TKqpSink& sink, TKqpTas
 }
 
 void TKqpTasksGraph::FillKqpTableSinkSettings(NKikimrKqp::TKqpTableSinkSettings& settings, const std::vector<std::pair<ui64, i64>>& internalSinksOrder, const TKqpTasksGraph::TTaskType& task) const {
-    auto& lockTxId = GetMeta().LockTxId;
+    const auto& lockTxId = GetMeta().LockTxId;
     if (lockTxId) {
         settings.SetLockTxId(*lockTxId);
         settings.SetLockNodeId(GetMeta().ExecuterId.NodeId());
@@ -2933,7 +1836,7 @@ void TKqpTasksGraph::FillKqpTableSinkSettings(NKikimrKqp::TKqpTableSinkSettings&
             }
         }
 
-    auto sinkPosition = std::lower_bound(
+    const auto* sinkPosition = std::lower_bound(
         internalSinksOrder.begin(),
         internalSinksOrder.end(),
         std::make_pair(task.StageId.TxId, settings.GetPriority()));
@@ -3031,88 +1934,21 @@ void TKqpTasksGraph::ResolveShards(TGraphMeta::TShardToNodeMap&& shardsToNodes) 
 size_t TKqpTasksGraph::BuildAllTasks(std::optional<TLlvmSettings> llvmSettings,
     const TVector<NKikimrKqp::TKqpNodeResources>& resourcesSnapshot, TQueryExecutionStats* stats)
 {
-    size_t sourceScanPartitionsCount = 0;
-
-    bool limitTasksPerNode = IsEnabledReadsMerge();
-
-    if (!GetMeta().IsScan) {
-        limitTasksPerNode |= GetMeta().StreamResult;
-    }
+    YQL_ENSURE(GetMeta().ShardsResolved);
 
     const auto internalSinksOrder = BuildInternalSinksPriorityOrder();
+    const auto sourceScanPartitionsCount = DoBuildAllTasks(resourcesSnapshot, stats);
 
     for (ui32 txIdx = 0; txIdx < Transactions.size(); ++txIdx) {
         const auto& tx = Transactions.at(txIdx);
-        auto scheduledTaskCount = ScheduleByCost(tx, resourcesSnapshot);
         for (ui32 stageIdx = 0; stageIdx < tx.Body->StagesSize(); ++stageIdx) {
             const auto& stage = tx.Body->GetStages(stageIdx);
             auto& stageInfo = GetStageInfo(NYql::NDq::TStageId(txIdx, stageIdx));
 
-            const bool maybeOlapRead = (GetMeta().AllowOlapDataQuery || GetMeta().StreamResult) && stageInfo.Meta.IsOlap();
-
-            // build task conditions
-            const bool buildFromSourceTasks = stage.SourcesSize() > 0;
-            const bool buildSysViewTasks = stageInfo.Meta.IsSysView();
-            const bool buildComputeTasks = stageInfo.Meta.ShardOperations.empty() || (!GetMeta().IsScan && stage.SinksSize() + stage.OutputTransformsSize() > 0 && !(maybeOlapRead && stageInfo.Meta.HasReads()));
-            const bool buildScanTasks = GetMeta().IsScan
-                ? stageInfo.Meta.IsOlap() || stageInfo.Meta.IsDatashard()
-                : maybeOlapRead && (stage.SinksSize() + stage.OutputTransformsSize() == 0 || stageInfo.Meta.HasReads())
-                ;
-
-            // TODO: doesn't work right now - multiple conditions are possible in tests
-            // YQL_ENSURE(buildFromSourceTasks + buildSysViewTasks + buildComputeTasks + buildScanTasks <= 1,
-            //     "Multiple task conditions: " <<
-            //     "isScan = " << (isScan) << ", "
-            //     "stage.SourcesSize() > 0 = " << (stage.SourcesSize() > 0) << ", "
-            //     "stageInfo.Meta.IsSysView() = " << (stageInfo.Meta.IsSysView()) << ", "
-            //     "stageInfo.Meta.ShardOperations.empty() = " << (stageInfo.Meta.ShardOperations.empty()) << ", "
-            //     "stage.SinksSize() = " << (stage.SinksSize()) << ", "
-            //     "stageInfo.Meta.IsOlap() = " << (stageInfo.Meta.IsOlap()) << ", "
-            //     "stageInfo.Meta.IsDatashard() = " << (stageInfo.Meta.IsDatashard()) << ", "
-            //     "AllowOlapDataQuery = " << (AllowOlapDataQuery) << ", "
-            //     "StreamResult = " << (StreamResult)
-            // );
-
             LOG_D("Stage " << stageInfo.Id << " AST: " << stage.GetProgramAst());
 
-            if (buildFromSourceTasks) {
-                switch (stage.GetSources(0).GetTypeCase()) {
-                    case NKqpProto::TKqpSource::kReadRangesSource: {
-                        // TODO: is this case only for reading OLTP table with datashards?
-                        if (auto partitionsCount = BuildScanTasksFromSource(stageInfo, limitTasksPerNode, stats)) {
-                            sourceScanPartitionsCount += *partitionsCount;
-                        } else {
-                            GetMeta().UnknownAffectedShardCount = true;
-                        }
-                    } break;
-                    case NKqpProto::TKqpSource::kFullTextSource: {
-                        BuildFullTextScanTasksFromSource(stageInfo, stats);
-                        GetMeta().UnknownAffectedShardCount = true;
-                    } break;
-                    case NKqpProto::TKqpSource::kSysViewSource: {
-                        BuildSysViewTasksFromSource(stageInfo);
-                        GetMeta().UnknownAffectedShardCount = true;
-                    } break;
-                    case NKqpProto::TKqpSource::kExternalSource: {
-                        YQL_ENSURE(!GetMeta().IsScan);
-                        auto it = scheduledTaskCount.find(stageIdx);
-                        BuildReadTasksFromSource(stageInfo, resourcesSnapshot, it != scheduledTaskCount.end() ? it->second.TaskCount : 0);
-                    } break;
-                    default:
-                        YQL_ENSURE(false, "unknown source type");
-                }
-            } else if (buildSysViewTasks) {
-                BuildSysViewScanTasks(stageInfo);
-            } else if (buildComputeTasks) {
-                auto nodesCount = GetMeta().ShardsOnNode.size();
-                if (!GetMeta().IsScan) {
-                    nodesCount = std::max<ui32>(resourcesSnapshot.size(), nodesCount);
-                }
-                GetMeta().UnknownAffectedShardCount |= BuildComputeTasks(stageInfo, nodesCount);
-            } else if (buildScanTasks) {
-                BuildScanTasksFromShards(stageInfo, tx.Body->EnableShuffleElimination(), CollectProfileStats(GetMeta().StatsMode) ? stats : nullptr);
-            } else {
-                AFL_ENSURE(false);
+            if (stage.GetIsSinglePartition()) {
+                YQL_ENSURE(stageInfo.Tasks.size() <= 1, "Unexpected multiple tasks in single-partition stage");
             }
 
             if (llvmSettings) {
@@ -3123,10 +1959,6 @@ size_t TKqpTasksGraph::BuildAllTasks(std::optional<TLlvmSettings> llvmSettings,
                 if (CollectProfileStats(GetMeta().StatsMode) && stats) {
                     stats->SetUseLlvm(stageInfo.Id.StageId, useLlvm);
                 }
-            }
-
-            if (stage.GetIsSinglePartition()) {
-                YQL_ENSURE(stageInfo.Tasks.size() <= 1, "Unexpected multiple tasks in single-partition stage");
             }
 
             for (const auto& taskId : stageInfo.Tasks) {
@@ -3151,7 +1983,7 @@ size_t TKqpTasksGraph::BuildAllTasks(std::optional<TLlvmSettings> llvmSettings,
 
 void TKqpTasksGraph::BuildLiteralTasks() {
     for (ui32 txIdx = 0; txIdx < Transactions.size(); ++txIdx) {
-        auto& tx = Transactions.at(txIdx);
+        const auto& tx = Transactions.at(txIdx);
 
         for (ui32 stageIdx = 0; stageIdx < tx.Body->StagesSize(); ++stageIdx) {
             auto& stageInfo = GetStageInfo(TStageId(txIdx, stageIdx));
@@ -3166,6 +1998,38 @@ void TKqpTasksGraph::BuildLiteralTasks() {
     }
 }
 
+std::optional<NYql::TIssue> TKqpTasksGraph::ValidateTasks() const {
+    auto ValidateChannel = [&](ui64 channelId) {
+        const auto& channel = GetChannel(channelId);
+
+        YQL_ENSURE(channel.SrcTask);
+        if (!GetMeta().AllowWithSpilling) {
+            YQL_ENSURE(channel.InMemory, "With spilling off, all channels should be stored in memory only. "
+                << "Not InMemory channelId: " << channelId);
+        }
+    };
+
+    try {
+        for (const auto& task : GetTasks()) {
+            for (const auto& input : task.Inputs) {
+                for (ui64 channelId : input.Channels) {
+                    ValidateChannel(channelId);
+                }
+            }
+
+            for (const auto& output : task.Outputs) {
+                for (ui64 channelId : output.Channels) {
+                    ValidateChannel(channelId);
+                }
+            }
+        }
+    } catch (const TYqlPanic& e) {
+        return YqlIssue({}, TIssuesIds::DEFAULT_ERROR, e.what());
+    }
+
+    return {};
+}
+
 TKqpTasksGraph::TKqpTasksGraph(
     const TString& database,
     const TVector<IKqpGateway::TPhysicalTxData>& transactions,
@@ -3176,10 +2040,10 @@ TKqpTasksGraph::TKqpTasksGraph(
     TIntrusiveConstPtr<NACLib::TUserToken> userToken)
     : Transactions(transactions)
     , TxAlloc(txAlloc)
+    , UserToken(std::move(userToken))
     , AggregationSettings(aggregationSettings)
     , Counters(counters)
     , BufferActorId(bufferActorId)
-    , UserToken(std::move(userToken))
 {
     GetMeta().Arena = MakeIntrusive<NActors::TProtoArenaHolder>();
     GetMeta().Database = database;
@@ -3247,7 +2111,7 @@ std::vector<std::pair<ui64, i64>> TKqpTasksGraph::BuildInternalSinksPriorityOrde
         const auto& tx = Transactions.at(txIdx);
         for (ui32 stageIdx = 0; stageIdx < tx.Body->StagesSize(); ++stageIdx) {
             const auto& stage = tx.Body->GetStages(stageIdx);
-            auto& stageInfo = GetStageInfo(NYql::NDq::TStageId(txIdx, stageIdx));
+            const auto& stageInfo = GetStageInfo(NYql::NDq::TStageId(txIdx, stageIdx));
 
             auto addSink = [&stageInfo, &order, txIdx](const NKqpProto::TKqpInternalSink& intSink) {
                 AFL_ENSURE(intSink.GetSettings().Is<NKikimrKqp::TKqpTableSinkSettings>());
@@ -3407,15 +2271,198 @@ TString TKqpTasksGraph::DumpToString() const {
     return dump.Str();
 }
 
+// static
+std::map<ui32, TKqpTasksGraph::TStageScheduleInfo> TKqpTasksGraph::ScheduleByCost(const IKqpGateway::TPhysicalTxData& tx, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot) {
+    std::map<ui32, TStageScheduleInfo> result;
+    if (!resourceSnapshot.empty()) { // can't schedule w/o node count
+        // collect costs and schedule stages with external sources only
+        double totalCost = 0.0;
+        for (ui32 stageIdx = 0; stageIdx < tx.Body->StagesSize(); ++stageIdx) {
+            const auto& stage = tx.Body->GetStages(stageIdx);
+            if (stage.SourcesSize() > 0 && stage.GetSources(0).GetTypeCase() == NKqpProto::TKqpSource::kExternalSource) {
+                if (stage.GetStageCost() > 0.0 && stage.GetTaskCount() == 0) {
+                    totalCost += stage.GetStageCost();
+                    result.emplace(stageIdx, TStageScheduleInfo{.StageCost = stage.GetStageCost()});
+                }
+            }
+        }
+        // assign task counts
+        if (!result.empty()) {
+            // allow use 2/3 of threads in single stage
+            ui32 maxStageTaskCount = ((TStagePredictor::GetUsableThreads() * 2) + 2) / 3;
+            // total limit per mode is x2
+            ui32 maxTotalTaskCount = maxStageTaskCount * 2;
+            for (auto& [_, stageInfo] : result) {
+                // schedule tasks evenly between nodes
+                stageInfo.TaskCount =
+                    std::max<ui32>(
+                        std::min(static_cast<ui32>(maxTotalTaskCount * stageInfo.StageCost / totalCost), maxStageTaskCount)
+                        , 1
+                    ) * resourceSnapshot.size();
+            }
+        }
+    }
+    return result;
+}
+
+// static
+void TKqpTasksGraph::FillReadTaskFromSource(TTask& task, const TString& sourceName, const TString& structuredToken, const TVector<NKikimrKqp::TKqpNodeResources>& resourceSnapshot, ui64 nodeOffset) {
+    if (structuredToken) {
+        task.Meta.SecureParams.emplace(sourceName, structuredToken);
+    }
+
+    if (resourceSnapshot.empty()) {
+        task.Meta.Type = TTaskMeta::ETaskType::Compute;
+    } else {
+        task.Meta.NodeId = resourceSnapshot[nodeOffset % resourceSnapshot.size()].GetNodeId();
+        task.Meta.Type = TTaskMeta::ETaskType::Scan;
+    }
+}
+
+// static
+TVector<TTaskMeta::TColumn> TKqpTasksGraph::BuildKqpColumns(const NKqpProto::TKqpReadRangesSource& op, TIntrusiveConstPtr<TTableConstInfo> tableInfo) {
+    return BuildKqpColumnsImpl(op, tableInfo);
+}
+
+// static
+TVector<TTaskMeta::TColumn> TKqpTasksGraph::BuildKqpColumns(const NKqpProto::TKqpPhyTableOperation& op, TIntrusiveConstPtr<TTableConstInfo> tableInfo) {
+    return BuildKqpColumnsImpl(op, tableInfo);
+}
+
+// static
+void TKqpTasksGraph::MergeReadInfoToTaskMeta(TTaskMeta& meta, ui64 shardId, TMaybe<TShardKeyRanges>& keyReadRanges,
+    const TPhysicalShardReadSettings& readSettings, const TVector<TTaskMeta::TColumn>& columns,
+    const NKqpProto::TKqpPhyTableOperation& op, bool isPersistentScan)
+{
+    TTaskMeta::TShardReadInfo readInfo = {
+        .Ranges = {},
+        .Columns = columns,
+    };
+    if (keyReadRanges) {
+        readInfo.Ranges = std::move(*keyReadRanges); // sorted & non-intersecting
+    }
+
+    if (isPersistentScan) {
+        readInfo.ShardId = shardId;
+    }
+
+    if (meta.Reads && !meta.Reads.GetRef().empty()) {
+        // Validate parameters
+        YQL_ENSURE(meta.ReadInfo.ItemsLimit == readSettings.ItemsLimit);
+        YQL_ENSURE(meta.ReadInfo.GetSorting() == readSettings.GetSorting());
+        // TODO: why no check for ReadType?
+    } else {
+        meta.ReadInfo.ItemsLimit = readSettings.ItemsLimit;
+        meta.ReadInfo.SetSorting(readSettings.GetSorting());
+        meta.ReadInfo.ReadType = TTaskMeta::TReadInfo::EReadType::Rows;
+    }
+
+    if (op.GetTypeCase() == NKqpProto::TKqpPhyTableOperation::kReadOlapRange) {
+        [&taskMeta = meta, resultType = readSettings.ResultType, &readOlapRange = std::as_const(op.GetReadOlapRange())] {
+            if (taskMeta.Reads && !taskMeta.Reads.GetRef().empty()) {
+                // Validate parameters
+                if (readOlapRange.GetOlapProgram().empty()) {
+                    YQL_ENSURE(taskMeta.ReadInfo.OlapProgram.Program.empty());
+                    return;
+                }
+
+                YQL_ENSURE(taskMeta.ReadInfo.OlapProgram.Program == readOlapRange.GetOlapProgram());
+                return;
+            }
+
+            if (resultType) {
+                YQL_ENSURE(resultType->GetKind() == NKikimr::NMiniKQL::TType::EKind::Struct
+                    || resultType->GetKind() == NKikimr::NMiniKQL::TType::EKind::Tuple);
+
+                auto* resultStructType = static_cast<NKikimr::NMiniKQL::TStructType*>(resultType);
+                ui32 resultColsCount = resultStructType->GetMembersCount();
+
+                taskMeta.ReadInfo.ResultColumnsTypes.reserve(resultColsCount);
+                for (ui32 i = 0; i < resultColsCount; ++i) {
+                    taskMeta.ReadInfo.ResultColumnsTypes.emplace_back();
+                    auto* memberType = resultStructType->GetMemberType(i);
+                    NScheme::TTypeInfo typeInfo = NScheme::TypeInfoFromMiniKQLType(memberType);
+                    taskMeta.ReadInfo.ResultColumnsTypes.back() = typeInfo;
+                }
+            }
+
+            if (readOlapRange.GetOlapProgram().empty()) {
+                return;
+            }
+
+            {
+                Y_ABORT_UNLESS(taskMeta.ReadInfo.GroupByColumnNames.empty());
+                std::vector<std::string> groupByColumns;
+                for (auto&& i : readOlapRange.GetGroupByColumnNames()) {
+                    groupByColumns.emplace_back(i);
+                }
+                std::swap(taskMeta.ReadInfo.GroupByColumnNames, groupByColumns);
+            }
+
+            switch (readOlapRange.GetReadType()) {
+                case NKqpProto::TKqpPhyOpReadOlapRanges::ROWS:
+                    taskMeta.ReadInfo.ReadType = TTaskMeta::TReadInfo::EReadType::Rows;
+                    break;
+                case NKqpProto::TKqpPhyOpReadOlapRanges::BLOCKS:
+                    taskMeta.ReadInfo.ReadType = TTaskMeta::TReadInfo::EReadType::Blocks;
+                    break;
+                default:
+                    YQL_ENSURE(false, "Invalid read type from TKqpPhyOpReadOlapRanges protobuf.");
+            }
+
+            taskMeta.ReadInfo.OlapProgram.Program = readOlapRange.GetOlapProgram();
+            for (const auto& name: readOlapRange.GetOlapProgramParameterNames()) {
+                taskMeta.ReadInfo.OlapProgram.ParameterNames.insert(name);
+            }
+        }();
+    }
+
+    if (!meta.Reads) {
+        meta.Reads.ConstructInPlace();
+    }
+
+    meta.Reads->emplace_back(std::move(readInfo));
+}
+
+// static
+void TKqpTasksGraph::PrepareScanMetaForUsage(TTaskMeta& meta, const TVector<NScheme::TTypeInfo>& keyTypes) {
+    YQL_ENSURE(meta.Reads.Defined());
+    auto& taskReads = meta.Reads.GetRef();
+
+    /*
+     * Sort read ranges so that sequential scan of that ranges produce sorted result.
+     *
+     * Partition pruner feed us with set of non-intersecting ranges with filled right boundary.
+     * So we may sort ranges based solely on the their rightmost point.
+     */
+    std::sort(taskReads.begin(), taskReads.end(), [&](const auto& lhs, const auto& rhs) {
+        if (lhs.ShardId == rhs.ShardId) {
+            return false;
+        }
+
+        const std::pair<const TSerializedCellVec*, bool> k1 = lhs.Ranges.GetRightBorder();
+        const std::pair<const TSerializedCellVec*, bool> k2 = rhs.Ranges.GetRightBorder();
+
+        const int cmp = CompareBorders<false, false>(
+            k1.first->GetCells(),
+            k2.first->GetCells(),
+            k1.second,
+            k2.second,
+            keyTypes);
+
+        return (cmp < 0);
+        });
+}
+
 TString TTaskMeta::ToString(const TVector<NScheme::TTypeInfo>& keyTypes, const NScheme::TTypeRegistry& typeRegistry) const {
     TStringBuilder sb;
     sb << "TTaskMeta{ ShardId: " << ShardId << ", Reads: { ";
 
     if (Reads) {
         for (ui64 i = 0; i < Reads->size(); ++i) {
-            auto& read = (*Reads)[i];
+            const auto& read = (*Reads)[i];
             sb << "[" << i << "]: { columns: [";
-            for (auto& x : read.Columns) {
+            for (const auto& x : read.Columns) {
                 sb << x.Name << ", ";
             }
             sb << "], ranges: " << read.Ranges.ToString(keyTypes, typeRegistry) << " }";
